requirements.txt
asgiref
Django
djangorestframework
mysqlclient
stripe